% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_network.R
\name{.kernel_capital_curly_f}
\alias{.kernel_capital_curly_f}
\title{F(tau) function}
\usage{
.kernel_capital_curly_f(
  infection = c("SIR", "SEIR"),
  transmission_rate,
  infectiousness_rate = NULL,
  recovery_rate,
  tau
)
}
\arguments{
\item{infection}{}
}
\description{
F(tau) function
}
