% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_network.R
\name{.kernel_capital_curly_f_derivative}
\alias{.kernel_capital_curly_f_derivative}
\title{Derivative F'(tau)}
\usage{
.kernel_capital_curly_f_derivative(
  infection = c("SIR", "SEIR"),
  transmission_rate,
  infectiousness_rate = NULL,
  recovery_rate,
  tau
)
}
\arguments{
\item{infection}{}
}
\description{
Derivative F'(tau)
}
