% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_network.R
\name{.kernel_captial_curly_f_inf_limit}
\alias{.kernel_captial_curly_f_inf_limit}
\title{Title}
\usage{
.kernel_captial_curly_f_inf_limit(
  infection = c("SIR", "SEIR"),
  transmission_rate,
  infectiousness_rate = NULL,
  recovery_rate
)
}
\arguments{
\item{recovery_rate}{}
}
\description{
Title
}
