% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_network.R
\name{helper_g_function}
\alias{helper_g_function}
\title{G'(x) / G'(1) = G'(x) / \if{html}{\out{<k>}}, \if{html}{\out{<k>}} = mean degree, G(x) probability generating function}
\usage{
helper_g_function(x, ...)
}
\arguments{
\item{...}{}
}
\description{
G'(x) / G'(1) = G'(x) / \if{html}{\out{<k>}}, \if{html}{\out{<k>}} = mean degree, G(x) probability generating function
}
