% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_network.R
\name{helper_g_function_derivative}
\alias{helper_g_function_derivative}
\title{Derivative of \link{helper_g_function}
G''(x) / G'(1) = G''(x) / \if{html}{\out{<k>}}, \if{html}{\out{<k>}} = mean degree, G(x) probability generating function}
\usage{
helper_g_function_derivative(x, ...)
}
\arguments{
\item{...}{}
}
\description{
Derivative of \link{helper_g_function}
G''(x) / G'(1) = G''(x) / \if{html}{\out{<k>}}, \if{html}{\out{<k>}} = mean degree, G(x) probability generating function
}
