% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_network.R
\name{model_network_seir}
\alias{model_network_seir}
\title{Title}
\usage{
model_network_seir(
  degree_distribution = c("poisson", "negative_binomial", "constant", "geometric"),
  transmission_rate = 0.25,
  infectiousness_rate = 0.1,
  recovery_rate = 0.25,
  population_size = 2e+07,
  seed_infected = 0.001,
  ...
)
}
\arguments{
\item{...}{}
}
\description{
Title
}
