% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_reference.R
\name{model_reference}
\alias{model_reference}
\title{Title}
\usage{
model_reference(
  transmission_rate = 0.25,
  infectiousness_rate = 0.25,
  recovery_rate = 0.2,
  time_end = 400,
  increment = 1,
  population_size = 2e+07,
  seed_infected = 0.001
)
}
\arguments{
\item{seed_infected}{}
}
\description{
Title
}
