% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models_combined.R
\name{models_combined}
\alias{models_combined}
\title{Title}
\usage{
models_combined(time_end, increment, params_p, params_nb, params_reference)
}
\arguments{
\item{params_reference}{}
}
\description{
Title
}
