% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_network.R
\name{simulate_outbreak_seir_network}
\alias{simulate_outbreak_seir_network}
\title{Title}
\usage{
simulate_outbreak_seir_network(t, increment, current_state, params)
}
\arguments{
\item{params}{}
}
\description{
Title
}
