% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_reference.R
\name{simulate_outbreak_seir_reference}
\alias{simulate_outbreak_seir_reference}
\title{Title}
\usage{
simulate_outbreak_seir_reference(t, increment, current_state, params)
}
\arguments{
\item{params}{}
}
\description{
Title
}
