% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epi_quantities_network.R
\name{theory_network_final_size}
\alias{theory_network_final_size}
\title{Title}
\usage{
theory_network_final_size(.params, .degree, .infection, .fraction = TRUE)
}
\arguments{
\item{.fraction}{}
}
\description{
Title
}
