% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epi_quantities_network.R
\name{theory_network_reproduction_number}
\alias{theory_network_reproduction_number}
\title{Title}
\usage{
theory_network_reproduction_number(.infection, .params, .degree)
}
\arguments{
\item{.degree}{}
}
\description{
Title
}
