% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epi_quantities_reference.R
\name{theory_reference_final_size}
\alias{theory_reference_final_size}
\title{Title}
\usage{
theory_reference_final_size(.params = NULL, r_0 = NULL, .fraction = TRUE)
}
\arguments{
\item{r_0}{}
}
\description{
Title
}
