% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epi_quantities_reference.R
\name{theory_reference_reproduction_number}
\alias{theory_reference_reproduction_number}
\title{Title}
\usage{
theory_reference_reproduction_number(.params)
}
\arguments{
\item{.params}{}
}
\description{
Title
}
